% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\docType{data}
\name{landmarks.2016}
\alias{landmarks.2016}
\alias{landmarks.2017}
\alias{landmarks}
\title{Landmarks for transformations between different assemblies of a refernce
brain.}
\description{
\code{landmarks.2016} landmarks set for 2016 brain

\code{landmarks.2017} landmarks set for 2017 brain

\code{landmarks} always refers to the latest landmark set
}
